fix(postgres): widen entry_revisions.revision_number and site_webhook_deliveries.duration_ms to BIGINT
Rust models declare these fields as i64 (sqlx BIGINT/INT8), but the
initial schema declared them as INTEGER (INT4). SQLite's type affinity
silently coerced INT4 to i64 on read, so the bug never surfaced. Postgres
is strict and rejected the decode with:
  'Rust type i64 (as SQL type INT8) is not compatible with SQL type INT4'

revision_number is monotonically increasing per entry and duration_ms
is unbounded, so BIGINT is the correct width for both.

+++ apps/backend/migrations/postgres/20260501000000_initial_schema.sql
@@ -113,7 +113,7 @@ CREATE INDEX IF NOT EXISTS idx_efr_entry ON entry_file_references(entry_id);
 CREATE TABLE IF NOT EXISTS entry_revisions (
     id TEXT PRIMARY KEY NOT NULL,
     entry_id TEXT NOT NULL REFERENCES entries(id) ON DELETE CASCADE,
-    revision_number INTEGER NOT NULL,
+    revision_number BIGINT NOT NULL,
     data JSONB NOT NULL,
     created_by TEXT REFERENCES users(id),
     created_at TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT NOW(),
@@ -142,7 +142,7 @@ CREATE TABLE IF NOT EXISTS site_webhook_deliveries (
     status VARCHAR(20) NOT NULL CHECK(status IN ('success', 'failed')),
     status_code INTEGER,
     response_body TEXT,
-    duration_ms INTEGER,
+    duration_ms BIGINT,
     triggered_by VARCHAR(36) REFERENCES users(id) ON DELETE SET NULL,
     triggered_at TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT NOW()
 );
